% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0tfABESSR.R
\name{BenchmarksABESS}
\alias{BenchmarksABESS}
\title{Benchmark l0tfABESSR and l0tfABESSC functions}
\description{
Use R package \code{microbenchmark} to compare the performance of R functions (\code{l0tfABESSR}) and RCpp functions (\code{l0tfABESSC}).
}
\examples{
\dontrun{
tm1 <- microbenchmark::microbenchmark(
  rnR = Simul0tfABESSR(sigma=0.1,dgm="Blocks",n=200),
  rnC = Simul0tfABESSC(sigma=0.1,dgm="Blocks",n=200),
  times = 100
)
print(summary(tm1)[,c(1,3,5,6)])
tm2 <- microbenchmark::microbenchmark(
  rnR = Simul0tfABESSR(sigma=0.1,dgm="Wave",n=200),
  rnC = Simul0tfABESSC(sigma=0.1,dgm="Wave",n=200),
  times = 100
)
print(summary(tm2)[,c(1,3,5,6)])
}
}
