% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0tfAMIASR.R
\name{DiffMat}
\alias{DiffMat}
\title{Generate a Difference Matrix based on AMIAS method}
\usage{
DiffMat(n, q)
}
\arguments{
\item{n}{The number of data points.}

\item{q}{The order of the difference to be computed.}
}
\value{
A matrix of size (n-q-1) by n, with elements arranged to facilitate the calculation of a 'q'-th order difference.
}
\description{
This function generates a matrix used for calculating differences of a certain order. It is particularly useful in numerical methods or for creating specific matrix patterns. The matrix created by this function can be used in various applications where a difference operation of order 'q' is needed over 'n' data points.
}
\examples{
\dontrun{
diffMatrix <- DiffMat(n = 10, q = 2)
print(diffMatrix)
}
}
