% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0tfABESSC.R
\name{l0tfABESSC}
\alias{l0tfABESSC}
\title{Sequential L0 Trend Filtering based on ABESS with Rcpp}
\usage{
l0tfABESSC(y = y, kmax = kmax, q = q)
}
\arguments{
\item{y}{Response data.}

\item{kmax}{Maximum number of detected knots for sequential trend filtering.}

\item{q}{Order and integer used in creating the differential matrix, influencing the filtering process.}
}
\value{
A list containing the optimal number of detected knots (kopt), the optimized beta coefficients, the full history of active sets (A.all), the final active set (A), and the predicted values (yhat) after the trend filtering process.
}
\description{
Given the response data, the maximum number of detected knots, and the order, we obtain the solution to the L0 trend filtering problem based on the idea of ABESS.
}
\examples{
\dontrun{
y <- rnorm(100)
result <- l0tfABESSC(y = y, kmax = 10, q = 1)
plot(y)
lines(result$yhat, col = "red")
}
}
