% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0tfAMIASR.R
\name{l0tfAMIASR}
\alias{l0tfAMIASR}
\title{L0 trend filtering using AMIAS method with R functions}
\usage{
l0tfAMIASR(y = y, kmax = kmax, q = q)
}
\arguments{
\item{y}{A numeric vector representing the data from which the model is to be identified.}

\item{kmax}{The maximum number of variables to be included in the model. The algorithm will stop when this number is reached.}

\item{q}{The order of the difference to be calculated, used within the 'DiffMat' function call inside.}
}
\value{
A list containing several components: 'kopt' - the optimal number of knots based on BIC, 'beta' - the model parameters at the optimal iteration, 'A.all' - a list containing the selected variables at each iteration, 'A' - the final set of selected variables at the optimal iteration.
}
\description{
This function implements an AMIAS algorithm using L0-norm minimization for sparse representations. The main objective of the function is to identify the most relevant variables in a dataset by minimizing the L0-norm. The function iteratively selects the most significant variables and computes the corresponding model parameters.
}
\examples{
\dontrun{
result <- l0tfAMIASR(y = rnorm(100), kmax = 10, q = 2)
print(result$kopt)
print(result$beta)

}
}
