% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0tfABESSR.R
\name{SimuBlocks}
\alias{SimuBlocks}
\title{Simulate 'Blocks' Data}
\usage{
SimuBlocks(n, sigma = 0.1, seed = NA)
}
\arguments{
\item{n}{Number of data points to generate.}

\item{sigma}{Standard deviation of the noise added to the signal.}

\item{seed}{An optional seed for random number generation to make results reproducible.}
}
\value{
A list containing the simulated data (y), the true signal without noise (y0), the x-coordinates (x), change points (tau), and indices of change points (SetA).
}
\description{
This function simulates data based on a 'blocks' signal, commonly used in benchmarking signal processing methods.
}
\examples{
\dontrun{
simData <- SimuBlocks(n = 100, sigma = 0.1)
plot(simData$x, simData$y)
lines(simData$x, simData$y0, col = "red")
}
}
