% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0tfABESSC.R
\name{Simul0tfABESSC}
\alias{Simul0tfABESSC}
\title{L0 trend filtering is calculated using the ABESS approach and Rcpp functions}
\usage{
Simul0tfABESSC(sigma = 0.1, dgm = "Blocks", n = n, seed = NA)
}
\arguments{
\item{sigma}{The standard deviation of the noise added to the simulated data.}

\item{dgm}{The type of data generation model to use. Possible values are "Blocks" or "Wave".}

\item{n}{The number of data points to be generated in the simulation.}

\item{seed}{An optional seed for random number generation to make the simulation reproducible.}
}
\value{
The result of the L0 trend filtering algorithm, typically including the optimal number of variables chosen, the coefficients of the model, and other relevant information.
}
\description{
Based on the concept of ABESS, the L0 trend filtering solution obtained using Rcpp functions estimates piecewise constant and piecewise linear functions.
}
\examples{
\dontrun{
# Simulating 'Blocks' data
blocksData <- SimuBlocks(n = 100, sigma = 0.1, seed=123)
resultBlocks <- Simul0tfABESSC(sigma = 0.1, dgm = "Blocks", n = 100, seed=123)
plot(blocksData$y0)
lines(resultBlocks$yhat,col="red")

# Simulating 'Wave' data
waveData <- SimuWave(n = 100, sigma = 0.1, seed=123)
resultWave <- Simul0tfABESSC(sigma = 0.1, dgm = "Wave", n = 100, seed=123)
plot(waveData$y0)
lines(resultWave$yhat,col="red")
}
}
