% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0tfABESSR.R
\name{Simul0tfABESSR}
\alias{Simul0tfABESSR}
\title{L0 Trend Filtering is calculated using the ABESS approach and R functions}
\usage{
Simul0tfABESSR(sigma = 0.1, dgm = "Blocks", n = n, seed = NA)
}
\arguments{
\item{sigma}{The standard deviation of the noise added to the simulated data.}

\item{dgm}{The type of data generation model to use. Possible values are "Blocks" or "Wave".}

\item{n}{The number of data points to be generated in the simulation.}

\item{seed}{An optional seed for random number generation to make the simulation reproducible.}
}
\value{
The result of the L0 trend filtering algorithm, typically including the optimal number of variables chosen, the coefficients of the model, and other relevant information.
}
\description{
This function applies L0 trend filtering to simulated data. It supports different types of simulated data, such as 'Blocks' and 'Wave'.
}
\examples{
\dontrun{
# Simulating 'Blocks' data
blocksData <- SimuBlocks(n = 100, sigma = 0.1, seed=123)
resultBlocks <- Simul0tfABESSR(sigma = 0.1, dgm = "Blocks", n = 100, seed=123)
plot(blocksData$y0)
lines(resultBlocks$yhat,col="red")

# Simulating 'Wave' data
waveData <- SimuWave(n = 100, sigma = 0.1, seed=123)
resultWave <- Simul0tfABESSR(sigma = 0.1, dgm = "Wave", n = 100, seed=123)
plot(waveData$y0)
lines(resultWave$yhat,col="red")
}
}
