% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0tfAMIASR.R
\name{Simul0tfAMIASR}
\alias{Simul0tfAMIASR}
\title{L0 trend filtering is calculated using the AMIAS approach and R functions}
\usage{
Simul0tfAMIASR(sigma = 0.1, dgm = "Blocks", n = n, seed = NA)
}
\arguments{
\item{sigma}{The standard deviation of the noise added to the simulated data.}

\item{dgm}{The type of data generation model to use. Possible values are "Blocks" or "Wave". These models determine the characteristics of the simulated data.}

\item{n}{The number of data points to be generated in the simulation.}

\item{seed}{An optional seed for random number generation to make the simulation reproducible.}
}
\value{
The result of applying the L0 trend filtering algorithm using AMIAS on the simulated data, typically including the optimal number of variables chosen, the coefficients of the model, and other relevant information.
}
\description{
This function applies L0 trend filtering to simulated data using an AMIAS approach and sparse representation. The function supports different types of simulated data generation models, such as 'Blocks' and 'Wave'.
}
\examples{
\dontrun{
# Simulating 'Blocks' data
blocksData <- SimuBlocks(n = 100, sigma = 0.1, seed=123)
resultBlocks <- Simul0tfAMIASR(sigma = 0.1, dgm = "Blocks", n = 100, seed=123)
plot(blocksData$y0)
lines(resultBlocks$yhat,col="red")

# Simulating 'Wave' data
waveData <- SimuWave(n = 100, sigma = 0.1, seed=123)
resultWave <- Simul0tfAMIASR(sigma = 0.1, dgm = "Wave", n = 100, seed=123)
plot(waveData$y0)
lines(resultWave$yhat,col="red")
}
}
