% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0tfABESSR.R
\name{SimuWave}
\alias{SimuWave}
\title{Simulate 'Wave' Data}
\usage{
SimuWave(n, sigma = 0.1, seed = NA)
}
\arguments{
\item{n}{Number of data points.}

\item{sigma}{Noise level.}

\item{seed}{Random seed for reproducibility.}
}
\value{
A list of the generated y values, the true signal (y0), x-coordinates (x), change points (tau), and indices of change points (SetA).
}
\description{
Generates data based on a 'wave' pattern. Useful for testing and benchmarking in signal processing and statistical analysis.
}
\examples{
\dontrun{
simWave <- SimuWave(n = 100, sigma = 0.1)
plot(simWave$x, simWave$y)
lines(simWave$x, simWave$y0, col = "blue")
}
}
